--- src/subdomains/core/payment-link/share/c2b-payment-status.ts
@@ -1,7 +1,0 @@
-export enum C2BPaymentStatus {
-    WAITING = 'WAITING',
-    PENDING = 'PENDING',
-    COMPLETED = 'COMPLETED',
-    FAILED = 'FAILED',
-    REFUNDED = 'REFUNDED',
-}
